refactor: 전화번호 중복 검사 문구 추가
